+++ CHANGELOG.md
@@ -1,6 +1,9 @@
-### upcoming
-* add config version
-* add method `/module/broadcast_event` to broadcast arbitrary event to all connected modules
+### v2.3.0
+* add method `/module/broadcast_event` to broadcast arbitrary event to modules
+* add module dependencies to response `/module/get_modules_info`
+* add option to create a new config version instead update it
+* fix broadcasting config on activating and upserting configs
+* update libs
 
 ### v2.2.1
 * update deps